now should work

+++ cleanup.py
@@ -24,8 +24,10 @@ def main(_):
         model_dir = os.path.join(FLAGS.bucket_name, "models", suggestion_id)
         param_config_file = os.path.join(FLAGS.bucket_name, "param_configs", "{}.json".format(suggestion_id))
 
-        tf.gfile.DeleteRecursively(model_dir)
-        tf.gfile.Remove(param_config_file)
+        if tf.gfile.exists(model_dir):
+            tf.gfile.DeleteRecursively(model_dir)
+        if tf.gfile.exists(param_config_file):
+            tf.gfile.Remove(param_config_file)
 
     conn.experiments(FLAGS.experiment_id).observations().delete(state="failed")
 
@@ -35,8 +37,10 @@ def main(_):
 
             model_dir = os.path.join(FLAGS.bucket_name, "models", suggestion.id)
             param_config_file = os.path.join(FLAGS.bucket_name, "param_configs", "{}.json".format(suggestion.id))
-            tf.gfile.DeleteRecursively(model_dir)
-            tf.gfile.Remove(param_config_file)
+            if tf.gfile.exists(model_dir):
+                tf.gfile.DeleteRecursively(model_dir)
+            if tf.gfile.exists(param_config_file):
+                tf.gfile.Remove(param_config_file)
         
         conn.experiments(FLAGS.experiment_id).suggestions().delete(state="open")
